+++ assets/images/onboarding-screen-illustrations/compass-illustration.png
