+++ assets/images/onboarding-screen-illustrations/prayer-illustration.png
