+++ src/features/onboarding/data/onboardingData.ts
@@ -1,0 +1,23 @@
+export const onboardingData = [
+  {
+    id: 1,
+    title: "Step-by-step Prayer Guide",
+    description: "Learn how to perform prayers correctly with detailed step-by-step guidance.",
+    imgsrc: require("@/assets/images/onboarding-screen-illustrations/prayer-illustration.png"),
+    imgPos: "top" as const,
+  },
+  {
+    id: 2,
+    title: "Daily and Monthly Prayer-timetable",
+    description: "Never miss a prayer with accurate daily and monthly prayer schedules.",
+    imgsrc: require("@/assets/images/onboarding-screen-illustrations/calender-illustration.png"),
+    imgPos: "middle" as const,
+  },
+  {
+    id: 3,
+    title: "Qibla Finder based on user's location",
+    description: "Find the correct prayer direction (Qibla) from anywhere in the world.",
+    imgsrc: require("@/assets/images/onboarding-screen-illustrations/compass-illustration.png"),
+    imgPos: "top" as const,
+  },
+];
